+++ packages/engine/src/avatar/AvatarSystem.ts
@@ -122,109 +122,5 @@ export default async function AvatarSystem(world: World) {
     // removeQuery(world, xrRGripQuery)
   }
 
-  return { execute, cleanup, subsystems: [() => import('./AvatarIKTargetSystem')] }
-}
-
-export function xrInputQueryExit(entity: Entity) {
-  removeComponent(entity, AvatarHeadIKComponent)
-  const { leftHint, rightHint } = getComponent(entity, AvatarHandsIKComponent)
-  leftHint?.removeFromParent()
-  rightHint?.removeFromParent()
-  removeComponent(entity, AvatarHandsIKComponent)
-  removeComponent(entity, AvatarArmsTwistCorrectionComponent)
-}
-
-/**
- * Setup head-ik for entity
- * @param entity
- * @returns
- */
-export function setupHeadIK(entity: Entity) {
-  // Add head IK Solver
-  const xrInputSourceComponent = getComponent(entity, XRInputSourceComponent)
-  addComponent(entity, AvatarHeadIKComponent, {
-    camera: xrInputSourceComponent.head,
-    rotationClamp: 0.785398
-  })
-}
-
-export function setupHandIK(entity: Entity) {
-  const xrInputSourceComponent = getComponent(entity, XRInputSourceComponent)
-
-  // Hands IK solver
-  const leftHint = new Object3D()
-  const rightHint = new Object3D()
-  const leftOffset = new Object3D()
-  const rightOffset = new Object3D()
-  const vec = new Vector3()
-
-  const animation = getComponent(entity, AvatarAnimationComponent)
-
-  leftOffset.rotation.set(-Math.PI * 0.5, Math.PI, 0)
-  rightOffset.rotation.set(-Math.PI * 0.5, 0, 0)
-
-  // todo: load the avatar & rig on the server
-  if (isClient) {
-    Object3DUtils.getWorldPosition(animation.rig.LeftShoulder, leftHint.position)
-    Object3DUtils.getWorldPosition(animation.rig.LeftArm, vec)
-    vec.subVectors(vec, leftHint.position).normalize()
-    leftHint.position.add(vec)
-    animation.rig.LeftShoulder.attach(leftHint)
-
-    Object3DUtils.getWorldPosition(animation.rig.RightShoulder, rightHint.position)
-    Object3DUtils.getWorldPosition(animation.rig.RightArm, vec)
-    vec.subVectors(vec, rightHint.position).normalize()
-    rightHint.position.add(vec)
-    animation.rig.RightShoulder.attach(rightHint)
-  }
-
-  addComponent(entity, AvatarHandsIKComponent, {
-    leftTarget: xrInputSourceComponent.controllerLeftParent,
-    leftHint: leftHint,
-    leftTargetOffset: leftOffset,
-    leftTargetPosWeight: 1,
-    leftTargetRotWeight: 1,
-    leftHintWeight: -1,
-    rightTarget: xrInputSourceComponent.controllerRightParent,
-    rightHint: rightHint,
-    rightTargetOffset: rightOffset,
-    rightTargetPosWeight: 1,
-    rightTargetRotWeight: 1,
-    rightHintWeight: -1
-  })
-
-  addComponent(entity, AvatarArmsTwistCorrectionComponent, {
-    LeftHandBindRotationInv: new Quaternion(),
-    LeftArmTwistAmount: 0.6,
-    RightHandBindRotationInv: new Quaternion(),
-    RightArmTwistAmount: 0.6
-  })
-}
-
-export function xrInputQueryEnter(entity: Entity) {
-  // if (isClient) initializeXRInputs(entity)
-
-  setupXRInputSourceContainer(entity)
-
-  if (!hasComponent(entity, NetworkObjectOwnedTag)) {
-    // proxifyXRInputs(entity)
-    setupHeadIK(entity)
-  }
-
-  setupHandIK(entity)
-}
-
-export function setupXRInputSourceContainer(entity: Entity) {
-  const xrInputSourceComponent = getComponent(entity, XRInputSourceComponent)
-
-  xrInputSourceComponent.container.add(
-    xrInputSourceComponent.controllerLeftParent,
-    xrInputSourceComponent.controllerGripLeftParent,
-    xrInputSourceComponent.controllerRightParent,
-    xrInputSourceComponent.controllerGripRightParent
-  )
-
-  xrInputSourceComponent.container.name = 'XR Container'
-  xrInputSourceComponent.head.name = 'XR Head'
-  Engine.instance.currentWorld.scene.add(xrInputSourceComponent.container, xrInputSourceComponent.head)
+  return { execute, cleanup, subsystems: [] }
 }
