+++ packages/engine/src/avatar/components/AvatarHeadIKComponent.ts
@@ -6,7 +6,7 @@ import { createMappedComponent } from '../../ecs/functions/ComponentFunctions'
  * Rotate the target bone with given camera
  */
 export type AvatarHeadIKComponentType = {
-  camera: Camera | Object3D
+  camera: Object3D
   /**
    * Clamp the angle between bone forward vector and camera forward in radians
    * Use 0 to disable
